requirements.txt
flask==2.3.3
apify-client==1.4.1
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
feedparser==6.0.10 